% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kleist}
\alias{kleist}
\alias{kleist_3d}
\title{kleist atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
kleist

kleist_3d
}
\description{
A historical Kleist functional segregation atlas based on files
released with  Pijnenburg et al., NeuroImage, 
239, 2021, (\href{https://doi.org/10.1016/j.neuroimage.2021.118274}{DOI})
Version 1; 15-01-2021.
}
\examples{
data(kleist)
data(kleist_3d)
}
\references{
Kleist K. (1934) Gehirnpathologie vornehmlich auf Grund der Kriegserfahrungen (Leipzig: Barth).
    
\itemize{
 \item{kleist}{ - kleist atlas}
 \item{kleist_3d}{ - kleist 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
