% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{kleist}
\alias{kleist}
\title{Kleist Atlas}
\usage{
kleist()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the historical Kleist 1934 functional segregation with
49 regions per hemisphere. Contains both 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\details{
Based on supplementary materials of Pijnenburg et al., NeuroImage,
239, 2021; Version 1; 15-01-2021.
}
\examples{
kleist()
}
\references{
Kleist K. (1934) Gehirnpathologie vornehmlich auf Grund der
Kriegserfahrungen (Leipzig: Barth).

\if{html}{\out{<div class="sourceCode">}}\preformatted{Pijnenburg R, et al. (2021) NeuroImage, 239, 118274.
\doi{10.1016/j.neuroimage.2021.118274}
}\if{html}{\out{</div>}}
}
\concept{ggseg_atlases}
